requirements.txt

Flask
Flask-Login
Flask-SQLAlchemy
Flask-Mail
Flask-Cors
Authlib
web3
requests
python-dotenv
